+++ RELEASE-NOTES.md
@@ -54,6 +54,8 @@
 - Rename 4do to opera (newer libretro 3do core)
 - Bump Theodore core (add emulation of Thomson TO7 and TO7/70 computers)
 - Set vice_x64sc as default c64 emulator + JiffyDOS support
+- Improve the management of external screens and selection from recalbox.conf by Chriskt78
+
 
 ### Fixes
 - Fix Odroid XU4 fan issue